requirements.txt
alembic==1.16.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
automaton==3.2.0
autopage==0.5.2
bcrypt==4.3.0
boto3==1.38.28
botocore==1.38.28
cachetools==6.0.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cliff==4.10.0
cmd2==2.6.0
cryptography==45.0.3
debtcollector==3.0.0
decorator==5.2.1
dnspython==2.7.0
dogpile.cache==1.4.0
eventlet==0.40.0
fastapi==0.115.12
fasteners==0.19
futurist==3.1.1
gnureadline==8.2.13
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
iso8601==2.1.0
Jinja2==3.1.6
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keystoneauth1==5.11.0
keystonemiddleware==10.10.0
kombu==5.5.4
lxml==5.4.0
Mako==1.3.10
mangum==0.19.0
MarkupSafe==3.0.2
msgpack==1.1.0
netaddr==1.3.0
networkx==3.5
openstacksdk==4.6.0
os-client-config==2.1.0
os-service-types==1.7.0
osc-lib==4.0.2
oslo.cache==3.11.0
oslo.concurrency==7.1.0
oslo.config==9.8.0
oslo.context==6.0.0
oslo.db==17.2.1
oslo.i18n==6.5.1
oslo.log==7.1.0
oslo.messaging==17.0.0
oslo.metrics==0.12.0
oslo.middleware==6.5.0
oslo.policy==4.5.1
oslo.reports==3.5.1
oslo.serialization==5.7.0
oslo.service==4.2.0
oslo.upgradecheck==2.5.0
oslo.utils==9.0.0
oslo.versionedobjects==3.7.0
packaging==24.2
Paste==3.10.1
PasteDeploy==3.1.0
pbr==6.1.1
pecan==1.6.0
pinecone==7.0.2
pinecone-plugin-assistant==1.6.1
pinecone-plugin-interface==0.0.7
platformdirs==4.3.8
pluggy==1.6.0
prettytable==3.16.0
prometheus_client==0.22.1
psutil==7.0.0
pycadf==4.0.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydot==4.0.0
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.26.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
PyPDF2==3.0.1
pyperclip==1.9.0
pytest==8.4.0
python-barbicanclient==7.1.0
python-cinderclient==9.7.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-glanceclient==4.8.0
python-heatclient==4.2.0
python-keystoneclient==5.6.0
python-multipart==0.0.20
python-neutronclient==11.5.0
python-novaclient==18.9.0
python-octaviaclient==3.11.0
python-openstackclient==8.1.0
python-swiftclient==4.8.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
repoze.lru==0.7
requests==2.32.3
requestsexceptions==1.4.0
rfc3986==2.0.0
Routes==2.5.1
rpds-py==0.25.1
s3transfer==0.13.0
simplegeneric==0.8.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
statsd==4.0.1
stevedore==5.4.1
taskflow==6.0.1
tenacity==9.1.2
testresources==2.0.2
testscenarios==0.5.0
testtools==2.7.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
vine==5.1.0
warlock==2.1.0
wcwidth==0.2.13
WebOb==1.8.9
Werkzeug==3.1.3
wrapt==1.17.2
WSME==0.12.1
yappi==1.6.10
zipp==3.22.0
python-jose[cryptography]
requests
